% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branchCorr_source.r
\name{Cor_xyFrom2}
\alias{Cor_xyFrom2}
\title{Model-based correlation from progenitor}
\usage{
Cor_xyFrom2(t, rates, progType1, progType2, type1, type2, progStructure)
}
\arguments{
\item{t}{The length of time}

\item{rates}{A rate vector containing self-renewal rates of first compartment, followed by differentiation rates of intermediate types,
death rates of intermediate types, differentiation rates of mature types, and death rates of mature types, i.e.
(renewHSC, diffProgA, diffProgB, ..., deathProgA, deathProgB, ... diffMature1, diffMature2, ..., deathMature1, deathMature2, ...)}

\item{progType1}{An index indicating the progenitor type that mature cell 1 can descend from}

\item{progType2}{An index indicating the progenitor type that mature cell 2 can descend from}

\item{type1}{An index indicating the type of mature cell 1}

\item{type2}{An index indicating the type of mature cell 2}

\item{progStructure}{A vector of length equal to number of mature types whose i'th entry contains the corresponding hidden progenitor type from which mature type i descends}
}
\value{
Correlation between specified mature type compartments at time t
}
\description{
This function computes the model-based correlation between two mature cell types given that the process begins with one progenitor.
The function is basically a wrapper around those used to compute variance and covariances.
}

