% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branchCorr_source.r
\name{expit}
\alias{expit}
\title{Expit function}
\usage{
expit(x)
}
\arguments{
\item{x}{A number or vector}
}
\value{
The value of expit(x)
}
\description{
Evaluates the expit function
}
\examples{
x = .25; expit(x)
x = 1:5; expit(x)
}

