% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branchCorr_source.r
\name{getObsCorr}
\alias{getObsCorr}
\title{Compute observed pairwise correlations}
\usage{
getObsCorr(data, obsTimes)
}
\arguments{
\item{data}{A matrix of mature cell type counts, whose rows correspond to independent realizations and are in the format produced by \code{\link{simCompObserved}} with vecOutput=TRUE}

\item{obsTimes}{A vector of corresponding observation times}
}
\value{
Matrix containing pairwise correlations at each observation time
}
\description{
Returns the empirical pairwise correlations between each pair of types, given a dataset of observed counts
and a vector containing the corresponding observation times. The data matrix should be in the same format
as returned by \code{\link{simCompObserved}} with vecOutput=TRUE
}

