% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branchCorr_source.r
\name{hyperGeoSample}
\alias{hyperGeoSample}
\title{Multivariate hypergeometric sampling of data}
\usage{
hyperGeoSample(data, sampSize)
}
\arguments{
\item{data}{A matrix containing independent realizations of discretely observed data from the model. Each row of this matrix is in the format returned by \code{\link{simCompObserved}} wiht vecOutput=T}

\item{sampSize}{The sample size or number of draws n in a hypergeometric distribution}
}
\description{
Samples hypergeometrically from a matrix containing discretely observed data from a stochastic compartmental model.
}

