% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branchCorr_source.r
\name{inferNLMINBSamplingCorr}
\alias{inferNLMINBSamplingCorr}
\title{Optimize correlation loss function}
\usage{
inferNLMINBSamplingCorr(initGuess, obsTimes, obsCorr, nSample, nTotal,
  trueDeaths, progStructure, max = 5000)
}
\arguments{
\item{initGuess}{Vector containing initial guess for par}

\item{obsTimes}{A vector of observation times}

\item{obsCorr}{Matrix of empirical correlations returned by \code{\link{getObsCorr}}}

\item{nSample}{A vector containing the sample sizes of each mature type as arguments to hypergeometric sampling; this is the number of draws n}

\item{nTotal}{A vector containing the total sizes of each mature population; this is the total population N in hypergeometric sampling}

\item{trueDeaths}{Vector containing the true death rates to be fixed}

\item{progStructure}{A vector of length equal to number of mature types whose i'th entry contains the corresponding hidden progenitor type from which mature type i descends}

\item{max}{Max iterations, default to 5000}
}
\value{
An \code{nlminb} object containing optimal parameters
}
\description{
This function uses the generic optimization package \code{nlminb} to optimize the objective function given by
\code{\link{samplingCorrObjective}}. Takes additional initial guess parameter and allows specification of true death rates,
which are often known and can be fixed throughout optimization.
}

