% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branchCorr_source.r
\name{M_2x}
\alias{M_2x}
\title{This function computes the model-based mean of a mature cell type compartment, given rates and that the process begins with one initial
progenitor.}
\usage{
M_2x(t, rates, progType, type, progStructure)
}
\arguments{
\item{t}{The length of time}

\item{rates}{A rate vector containing self-renewal rates of first compartment, followed by differentiation rates of intermediate types,
death rates of intermediate types, differentiation rates of mature types, and death rates of mature types, i.e.
(renewHSC, diffProgA, diffProgB, ..., deathProgA, deathProgB, ... diffMature1, diffMature2, ..., deathMature1, deathMature2, ...)}

\item{progType}{An index indicating the initial progenitor type}

\item{type}{An index indicating the mature cell type}

\item{progStructure}{A vector of length equal to number of mature types whose i'th entry contains the corresponding hidden progenitor type from which mature type i descends}
}
\value{
The mean population of compartment `type' after time t
}
\description{
This function computes the model-based mean of a mature cell type compartment, given rates and that the process begins with one initial
progenitor.
}

