% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branchCorr_source.r
\name{marginalizedCor_xy}
\alias{marginalizedCor_xy}
\title{Marginalized model-based correlations}
\usage{
marginalizedCor_xy(t, rates, initProb, type1, type2, progStructure)
}
\arguments{
\item{t}{The time length}

\item{rates}{A rate vector containing self-renewal rates of first compartment, followed by differentiation rates of intermediate types,
death rates of intermediate types, differentiation rates of mature types, and death rates of mature types, i.e.
(renewHSC, diffProgA, diffProgB, ..., deathProgA, deathProgB, ... diffMature1, diffMature2, ..., deathMature1, deathMature2, ...)}

\item{initProb}{vector containing initial probabilities of beginning in HSC comparmtent and each progenitor compartment}

\item{type1}{Index of mature type 1}

\item{type2}{Index of mature type 2}

\item{progStructure}{A vector of length equal to number of mature types whose i'th entry contains the corresponding hidden progenitor type from which mature type i descends}
}
\value{
The model-based marginalized correlation between types at time t
}
\description{
Computes the marginalized correlation between two mature cell types, given the initial probability distribution.
}

