% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branchCorr_source.r
\name{optimizeSamplingCorrNLMINB}
\alias{optimizeSamplingCorrNLMINB}
\title{Infer optimal rates via correlation loss function}
\usage{
optimizeSamplingCorrNLMINB(data, numInits, initMean, obsTimes, nSample, nTotal,
  trueDeaths, progStructure)
}
\arguments{
\item{data}{A matrix of observed counts in the same format as produced by \code{\link{simCompObserved}}}

\item{numInits}{Number of random restarts}

\item{initMean}{The mean for the random initial parameters}

\item{obsTimes}{A vector of observation times}

\item{nSample}{A vector containing the sample sizes of each mature type as arguments to hypergeometric sampling; this is the number of draws n}

\item{nTotal}{A vector containing the total sizes of each mature population; this is the total population N in hypergeometric sampling}

\item{trueDeaths}{Vector containing the true death rates to be fixed}

\item{progStructure}{A vector of length equal to number of mature types whose i'th entry contains the corresponding hidden progenitor type from which mature type i descends}
}
\value{
Vector containing the best fitting parameters
}
\description{
This function is a wrapper for \code{\link{inferNLMINBSamplingCorr}} that takes a dataset in the format produced by
\code{\link{simCompObserved}} and infers its most likely parameters. Computes the empirical correlations and optimizes
the correlation loss function over many random initializations, and returns the best estimate in terms of lowest objective
function value.
}

