% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branchCorr_source.r
\name{randInit}
\alias{randInit}
\title{Generate random initial population vector}
\usage{
randInit(initProbs, totalTypes)
}
\arguments{
\item{initProbs}{The initial distribution vector}

\item{totalTypes}{The total number of compartments or types}
}
\value{
A vector of zeros in all but one entry, indicating the initial type
}
\description{
This function converts an initial distribution to an initial population indicator vector. It samples from
the initial distribution, and returns an initial population vector with an indicator
of 1 in the compartment from which the process will begin, with zeros elsewhere.
}

