% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branchCorr_source.r
\name{samplingCor_xy}
\alias{samplingCor_xy}
\title{Marginalized model-based correlations with sampling}
\usage{
samplingCor_xy(t, rates, initProb, type1, type2, pop1, pop2, total1, total2,
  progStructure)
}
\arguments{
\item{t}{The time length}

\item{rates}{A rate vector containing self-renewal rates of first compartment, followed by differentiation rates of intermediate types,
death rates of intermediate types, differentiation rates of mature types, and death rates of mature types, i.e.
(renewHSC, diffProgA, diffProgB, ..., deathProgA, deathProgB, ... diffMature1, diffMature2, ..., deathMature1, deathMature2, ...)}

\item{initProb}{vector containing initial probabilities of beginning in HSC comparmtent and each progenitor compartment}

\item{type1}{Index of mature type 1}

\item{type2}{Index of mature type 2}

\item{pop1}{Sample size of mature type 1}

\item{pop2}{Sample size of mature type 2}

\item{total1}{Total population of mature type 1 cells}

\item{total2}{Total population of mature type 2 cells}

\item{progStructure}{A vector of length equal to number of mature types whose i'th entry contains the corresponding hidden progenitor type from which mature type i descends}
}
\value{
Correlation between types at time t after marginalizing over sampling distribution and initial distribution
}
\description{
Computes the marginalized correlation between two mature cell types, given the initial probability distribution, incorporating
the effect of hypergeometric sampling.
}

