% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branchCorr_source.r
\name{samplingCorrObjective}
\alias{samplingCorrObjective}
\title{Correlation loss function objective}
\usage{
samplingCorrObjective(par, obsTimes, obsCorr, nSample, nTotal, progStructure)
}
\arguments{
\item{par}{A vector containing the process rates followed by all but the first component of the initial distribution vector. The rates
should be ordered as before, beginning with self-renewal rates of first compartment, followed by differentiation rates of intermediate types,
death rates of intermediate types, differentiation rates of mature types, and death rates of mature types, i.e.
(renewHSC, diffProgA, diffProgB, ..., deathProgA, deathProgB, ... diffMature1, diffMature2, ..., deathMature1, deathMature2, ...)}

\item{obsTimes}{A vector of observation times}

\item{obsCorr}{Matrix of empirical correlations returned by \code{\link{getObsCorr}}}

\item{nSample}{A vector containing the sample sizes of each mature type as arguments to hypergeometric sampling; this is the number of draws n}

\item{nTotal}{A vector containing the total sizes of each mature population; this is the total population N in hypergeometric sampling}

\item{progStructure}{A vector of length equal to number of mature types whose i'th entry contains the corresponding hidden progenitor type from which mature type i descends}
}
\value{
Objective value
}
\description{
Computes the objective function of the correlation-matching loss function estimator. Takes in model parameters and a matrix
of empirical correlations in the format produced by \code{\link{getObsCorr}}. This can then be plugged into a generic optimization
routine such as \code{nlminb}. Natural constraints such as positivity of parameters are enforced via a log-barrier penalty.
}

