% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branchCorr_source.r
\name{sim.once}
\alias{sim.once}
\title{Simulate discretely observed data from a stochastic compartmental model with error check}
\usage{
sim.once(t.end, initPopulation, rates, progStructure, obsTimes,
  vecOutput = FALSE, maxEvents = 999999999999)
}
\arguments{
\item{t.end}{The total time until end of simulation}

\item{initPopulation}{A vector of length equal to total number of compartments/types; entries contain the initial population of each type}

\item{rates}{A rate vector containing self-renewal rates of first compartment, followed by differentiation rates of intermediate types,
death rates of intermediate types, differentiation rates of mature types, and death rates of mature types, i.e.
(renewHSC, diffProgA, diffProgB, ..., deathProgA, deathProgB, ... diffMature1, diffMature2, ..., deathMature1, deathMature2, ...)}

\item{progStructure}{A vector of length equal to number of mature types whose i'th entry contains the corresponding hidden progenitor type from which mature type i descends}

\item{obsTimes}{A vector containing the observation times}

\item{vecOutput}{Logical, whether to return a vectorized representation of the matrix of observation times by population counts}

\item{maxEvents}{The maximum number of events to simulate}
}
\description{
Wrapper for \code{\link{simCompObserved}} that checks for no error code
}

