% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branchCorr_source.r
\name{simCompartments}
\alias{simCompartments}
\title{Forward simulate from a stochastic compartmental model}
\usage{
simCompartments(t.end, initPopulation, rates, progStructure,
  maxEvents = 999999999999)
}
\arguments{
\item{t.end}{The total time until end of simulation}

\item{initPopulation}{A vector of length equal to total number of compartments/types; entries contain the initial population of each type}

\item{rates}{A rate vector containing self-renewal rates of first compartment, followed by differentiation rates of intermediate types,
death rates of intermediate types, differentiation rates of mature types, and death rates of mature types, i.e.
(renewHSC, diffProgA, diffProgB, ..., deathProgA, deathProgB, ... diffMature1, diffMature2, ..., deathMature1, deathMature2, ...)}

\item{progStructure}{A vector of length equal to number of mature types whose i'th entry contains the corresponding hidden progenitor type from which mature type i descends}

\item{maxEvents}{The maximum number of events to simulate}
}
\value{
A vector containing the mature type populations at end of simulation period (time t=t.end).
}
\description{
Uses Gillespie forward simulation from a stochastic compartmental model, i.e. representing a hematopoietic tree.
}
\examples{
progStructure <- c(1,1,1,2,2)
initPopulation <- c(10,0,0,0,0,0,0,0)
rates <- c(.3,.2,.5,.06, .03, 2, 4, 5, 3, 1, .15,.5,.8,.1,.05)
simCompartments(5, initPopulation, rates,progStructure)
}

