% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branchCorr_source.r
\name{U_xy}
\alias{U_xy}
\title{Second cross-moments starting from HSC}
\usage{
U_xy(t, rates, progType1, progType2, type1, type2, progStructure)
}
\arguments{
\item{t}{The length of time}

\item{rates}{A rate vector containing self-renewal rates of first compartment, followed by differentiation rates of intermediate types,
death rates of intermediate types, differentiation rates of mature types, and death rates of mature types, i.e.
(renewHSC, diffProgA, diffProgB, ..., deathProgA, deathProgB, ... diffMature1, diffMature2, ..., deathMature1, deathMature2, ...)}

\item{progType1}{An index indicating the progenitor type that mature cell 1 can descend from}

\item{progType2}{An index indicating the progenitor type that mature cell 2 can descend from}

\item{type1}{An index indicating the type of mature cell 1}

\item{type2}{An index indicating the type of mature cell 2}

\item{progStructure}{A vector of length equal to number of mature types whose i'th entry contains the corresponding hidden progenitor type from which mature type i descends}
}
\value{
Value of second cross-moment at time t
}
\description{
This function computes the model-based second cross-moments, denoted U_mn|0 in the manuscript, between two mature cell types,
given rates and that the process begins with one initial hematopoietic stem cell (compartment 1).
}

